### 2016-12-19 #145 @whedon
[PRE REVIEW]: eLabFTW - a free and open source electronic lab notebook
**Submitting author:** @NicolasCARPi (<a href="http://orcid.org/0000-0002-7494-5555">Nicolas CARPI</a>)
**Repository:** <a href="https://github.com/elabftw/elabftw" target ="_blank">https://github.com/elabftw/elabftw</a>
**Version:** 1.4.0
