### 2016-12-19 b9664b6
make todolist persistent in SQL, make it sortable too. Fix #124. WIP
