### 2016-12-19 #145 @cMadan
For anyone coming across this pre-review issue from Twitter or otherwise, feel free to make suggestions in the review thread (https://github.com/openjournals/joss-reviews/issues/146) or as issues in the project's repository (https://github.com/elabftw/elabftw). While we don't need more reviewers, more constructive feedback to the authors is always good! :)
