### 2016-12-19 #146 @raamana
Hey, @NicolasCARPi, you need to add a plain text LICENSE containing ones of the OSI approved licenses: https://opensource.org/licenses/alphabetical
