### 2016-12-19 #146 @raamana
To authors: I was able to follow all the steps outlined here at https://elabftw.readthedocs.io/en/latest/install-mac.html

and landed into an error at the final step.

See the image here:
