### 2016-12-19 #146 @raamana
The description of the repo could be improved by opening with the 3 questions listed under the GIF and then proceeding with what is currently the first para.

Under the Installation section, authors could summarize the major requirements, write a line or two about the process, rather than simply linking to another site. This helps the user get an overall idea of the level of involvement and effort needed before getting into details.
