### 2016-12-19 #146 @NicolasCARPi
Dear @raamana , thank you for your constructive criticism :).

1. The licence file was removed from the root to be integrated first in a doc/ folder, which was then moved to another repo. I realize now that this repo does not have a plaintext licence file (although it is in the [documentation](https://elabftw.readthedocs.io/en/latest/license.html). Let me fix that (it is now done on hypernext branch, I'll merge it to master later, if needed).

2. Indeed since the 1.4.0 release a few days ago, composer needs to be installed to populate the vendor/ directory. I updated the documentation for Mac. Commit here : https://github.com/elabftw/elabdoc/commit/45e5e3e3bd94103b91f37c7c6027c86c36428a5a
Doc here : https://elabftw.readthedocs.io/en/latest/install-mac.html

3. The 3 questions have been moved as suggested.

A line or two have been added under the installation section to describe better how it works and what it needs. New readme visible [here](https://github.com/elabftw/elabftw/tree/hypernext).

Cheers,
~Nico
