### 2016-12-21 #146 @raamana
Hi Nicolas,

The installation instructions for Composer do not work as noted on their website: https://getcomposer.org/download/

Perhaps it may be better package it with your codebase to reduce the burden on the user, if at all it is possible?
