### 2016-12-21 #146 @NicolasCARPi
Hello,

You are on a Mac right ? Try with http://.
See this issue https://github.com/composer/composer/issues/5685.

Also, composer should not be packaged with the software, it isn't a correct deployment strategy to do that. I used to have the `vendor` directory tracked by git, but I removed it to reduce the size of the repo.

Now you should know that the default and recommended way to install this software is with Docker, and the docker image takes care of everything, to highly "reduce the burden on the user", in your terms :). Installing it on a server is very quick and easy, as you can see on the [install page](https://elabftw.readthedocs.io/en/latest/install.html).

eLabFTW is a server software, and should be installed on a server. Because of its flexibility it is also possible to install it on a personnal computer (Mac or Windows), but it is not the philosophy.

This software aims for correct and secure cryptography, and this can rise issues on the Mac OS X platform because of its outdated OpenSSL, PHP,  versions and libraries. This is why the recommended operating system is GNU/Linux (and inside a Docker container).
