### 2016-12-22 #146 @raamana
That clarification helps. I tried to follow the instructions as I was led. What this means is that the user at the very first entry point was not led to the recommended way to install it. So I would recommend updating the instructions to install to follow your line of thought (what you recommend and why), and only unveil more progressively advanced options only as needed, and with appropriate advice.
