### 2016-12-22 #146 @NicolasCARPi
Well, there is a WARNING saying exactly that on top of the Mac and Windows install pages. It is also repeated several times that it needs to be installed on a server, in the Readme, on the documentation home page, on the main website. I don't see how I can make it clearer!

![a](http://i.imgur.com/Gx9BZkH.png)
