### 2016-12-22 5dee918
make todoitems editable
