+++ js/todolist.js
@@ -1,181 +1,0 @@
-/*
- * todolist.js
- * this file is minified in todolist.min.js, and this is what is used in the code
- * so if you change stuff here, don't forget to minify ;)
- * part of elabftw.net project
- * adapted from http://web.koesbong.com/2011/01/24/sortable-and-editable-to-do-list-using-html5s-localstorage/
- */
-
-$(function() {
-    /*
-     * FUNCTIONS
-     */
-
-    /*
-    function remove($this) {
-        var parentId = $this.parent().attr('id');
-        // Remove todo list from localStorage based on the id of the clicked parent element
-        localStorage.removeItem(
-            "'" + parentId + "'"
-        );
-        // Fade out the list item then remove from DOM
-        $this.parent().fadeOut(function() {
-            $this.parent().remove();
-            regenerate();
-        });
-    }
-
-    function regenerate() {
-        var $todoItemLi = $('#show-items li');
-        // Empty the order array
-        order.length = 0;
-
-        // Go through the list item, grab the ID then push into the array
-        $todoItemLi.each(function() {
-            var id = $(this).attr('id');
-            order.push(id);
-        });
-
-        // Convert the array into string and save to localStorage
-        localStorage.setItem(
-            'todo-orders', order.join(',')
-        );
-    }
-    /*
-    function clearAll() {
-        var $todoListLi = $('#show-items li');
-
-        order.length = 0;
-        localStorage.clear();
-        $todoListLi.remove();
-    }
-
-    /*
-     * VARS
-     */
-
-    var i = Number(localStorage.getItem('todo-counter')) + 1,
-        j = 0,
-        k,
-        $form = $('#todo-form'),
-        $removeLink = $('#show-items li a'),
-        $itemList = $('#show-items'),
-        $clearAll = $('#clear-all'),
-        $newTodo = $('#todo'),
-        order = [],
-        orderList;
-
-    /*
-     * START
-     */
-        /*
-
-    // Initial loading of todo items
-    orderList = localStorage.getItem('todo-orders');
-
-    orderList = orderList ? orderList.split(',') : [];
-
-
-    /*
-    // display items
-    for( j = 0, k = orderList.length; j < k; j++) {
-        $itemList.append(
-            "<li id='" + orderList[j] + "'>"
-            + localStorage.getItem(orderList[j])
-            + " <a href='#'>X</a></li>"
-        );
-    }
-    */
-
-
-    // Create
-    $form.submit(function(e) {
-        e.preventDefault();
-        var body = $newTodo.val();
-        var currentdate = new Date();
-        var datetime = currentdate.getFullYear() + "-" +
-            (currentdate.getMonth()+1)  + "-" +
-            currentdate.getDate() + " " +
-            currentdate.getHours() + ":" +
-            currentdate.getMinutes() + ":" +
-            currentdate.getSeconds();
-        if (body !== "") {
-            $.post("app/controllers/TodolistController.php", {
-                create: true,
-                body: body
-            }).done(function(data) {
-                var json = JSON.parse(data);
-                if (json.res) {
-                    $('#todoItems-list').append("<li class='todoItem' id='todoItem_" +
-                            json.id +
-                            "'><a href='#' onClick='destroyTodolist(" +
-                            json.id +
-                            ")'>X</a><span style='font-size:60%;display:block;'>" +
-                            datetime + "</span>" + body +
-                            '</li>');
-                } else {
-                    notif(json.msg, 'ko');
-                }
-            });
-            /*
-            // Take the value of the input field and save it to localStorage
-            localStorage.setItem(
-                "todo-" + i, $newTodo.val()
-            );
-
[…]
