### 2017-01-09 #146 @raamana
Thanks for the reminder, Chris. I will try to give it another to run and test the software on my Mac.

As @NicolasCARPi points out, it should be installed on a server, and ideally be tested by a team, neither of which I don't have access to, for this purpose. So wondering what would be an ideal way to proceed from here.
