### 2017-01-10 #146 @NicolasCARPi
@raamana I can set up a server for you if you wish. It only costs a few dollars per month :) Tell me and I'll give you an IP adress so you can follow this tutorial:
https://www.youtube.com/watch?v=K8CbQl7H5WA
