### 2017-01-12 1b0f03d
add possibility to decode asn1 tokens from elab. Fix #315
