### 2017-01-21 e18abe6
remove js/ folder and add all min.js in app/js
