+++ js/colorpicker/i18n/jquery.ui.colorpicker-en-GB.js
@@ -1,27 +1,0 @@
-;jQuery(function($) {
-	$.colorpicker.regional['en-GB'] = {
-		ok:				'OK',
-		cancel:			'Cancel',
-		none:			'None',
-		button:			'Colour',
-		title:			'Pick a colour',
-		transparent:	'Transparent',
-		hsvH:			'H',
-		hsvS:			'S',
-		hsvV:			'V',
-		rgbR:			'R',
-		rgbG:			'G',
-		rgbB:			'B',
-		labL:			'L',
-		labA:			'a',
-		labB:			'b',
-		hslH:			'H',
-		hslS:			'S',
-		hslL:			'L',
-		cmykC:			'C',
-		cmykM:			'M',
-		cmykY:			'Y',
-		cmykK:			'K',
-		alphaA:			'A'
-	};
-});
