### 2017-01-22 7e7f32d
add app/js to ignored files for codeclimate
