### 2017-01-25 44d44d6
remove go_url js function and code duplication
