### 2017-01-25 4899009
remove import_sql_structure()
