+++ Gruntfile.js
@@ -24,8 +24,8 @@ module.exports = function(grunt) {
               'bower_components/colorpicker/jquery.colorpicker.js',
               'bower_components/fancybox/source/jquery.fancybox.js',
               'bower_components/jeditable/jquery.jeditable.js',
-              'bower_components/jquery.complexify.js/jquery.complexify.js',
-              'bower_components/jquery.complexify.js/jquery.complexify.banlist.js',
+              'bower_components/jquery.complexify/jquery.complexify.js',
+              'bower_components/jquery.complexify/jquery.complexify.banlist.js',
               'bower_components/keymaster/keymaster.js',
               'app/js/common.js',
               'app/js/cornify.js',
