### 2017-01-25 adc2460
alpha version of doodle on experiments. Fix #198
