### 2017-01-25 40e8e70
bigger canvas and hide it by default
