### 2017-01-28 #146 @JohnGriffiths
@cmadan - I am on board and ready to start reviewing. But despite a lot of browsing this repo it's still not clear to me how I am supposed to do this. Sorry. Please advise.
