### 2017-01-29 fd2734c
add new mention delimiter: # for team XP and $ for user XP. fix #334
