+++ app/js/plugins/mention/plugin.js
@@ -1,166 +1,381 @@
-/*jslint forin: true */
+/*global tinymce, jQuery */
 
-;(function($) {
-    $.fn.extend({
-        mention: function(options) {
-            this.opts = {
-                users: [],
-                delimiter: '@',
-                sensitive: true,
-                emptyQuery: false,
-                queryBy: ['name', 'username'],
-                typeaheadOpts: {}
+(function (tinymce, $) {
+    'use strict';
+
+    var AutoComplete = function (ed, options) {
+        this.editor = ed;
+
+        this.options = $.extend({}, {
+            source: [],
+            delay: 500,
+            queryBy: 'name',
+            items: 10
+        }, options);
+
+        this.matcher = this.options.matcher || this.matcher;
+        this.renderDropdown = this.options.renderDropdown || this.renderDropdown;
+        this.render = this.options.render || this.render;
+        this.insert = this.options.insert || this.insert;
+        this.highlighter = this.options.highlighter || this.highlighter;
+
+        this.query = '';
+        this.hasFocus = true;
+
+        this.renderInput();
+
+        this.bindEvents();
+    };
+
+    AutoComplete.prototype = {
+
+        constructor: AutoComplete,
+
+        renderInput: function () {
+            var rawHtml =  '<span id="autocomplete">' +
+                                '<span id="autocomplete-delimiter">' + this.options.delimiter + '</span>' +
+                                '<span id="autocomplete-searchtext"><span class="dummy">\uFEFF</span></span>' +
+                            '</span>';
+
+            this.editor.execCommand('mceInsertContent', false, rawHtml);
+            this.editor.focus();
+            this.editor.selection.select(this.editor.selection.dom.select('span#autocomplete-searchtext span')[0]);
+            this.editor.selection.collapse(0);
+        },
+
+        bindEvents: function () {
+            this.editor.on('keyup', this.editorKeyUpProxy = $.proxy(this.rteKeyUp, this));
+            this.editor.on('keydown', this.editorKeyDownProxy = $.proxy(this.rteKeyDown, this), true);
+            this.editor.on('click', this.editorClickProxy = $.proxy(this.rteClicked, this));
+
+            $('body').on('click', this.bodyClickProxy = $.proxy(this.rteLostFocus, this));
+
+            $(this.editor.getWin()).on('scroll', this.rteScroll = $.proxy(function () { this.cleanUp(true); }, this));
+        },
+
+        unbindEvents: function () {
+            this.editor.off('keyup', this.editorKeyUpProxy);
+            this.editor.off('keydown', this.editorKeyDownProxy);
+            this.editor.off('click', this.editorClickProxy);
+
+            $('body').off('click', this.bodyClickProxy);
+
+            $(this.editor.getWin()).off('scroll', this.rteScroll);
+        },
+
+        rteKeyUp: function (e) {
+            switch (e.which || e.keyCode) {
+            //DOWN ARROW
+            case 40:
+            //UP ARROW
+            case 38:
+            //SHIFT
+            case 16:
+            //CTRL
+            case 17:
+            //ALT
+            case 18:
+                break;
+
+            //BACKSPACE
+            case 8:
+                if (this.query === '') {
+                    this.cleanUp(true);
+                } else {
+                    this.lookup();
+                }
+                break;
+
+            //TAB
+            case 9:
+            //ENTER
+            case 13:
+                var item = (this.$dropdown !== undefined) ? this.$dropdown.find('li.active') : [];
+                if (item.length) {
+                    this.select(item.data());
+                    this.cleanUp(false);
+                } else {
+                    this.cleanUp(true);
+                }
+                break;
+
+            //ESC
+            case 27:
+                this.cleanUp(true);
+                break;
+
[…]
