### 2017-01-30 #146 @cMadan
@JohnGriffiths - For the most part you need to see if the statements above in the first post are true and check things off accordingly. If something isn't the case, or you otherwise have a suggestion, either post a comment to @NicolasCARPi here (or ask me) or make an issue in the project's repository (https://github.com/elabftw/elabftw) and we'll sort things out until the submission meets JOSS' guidelines. The guidelines are explained in a bit more detail here: http://joss.theoj.org/about#reviewer_guidelines.

@raamana - the current JOSS submission system only allows one reviewer to be assigned at a time, so I'm going to change it to @JohnGriffiths. Nonetheless, please feel free to continue to open issues in the project's repository and make comments here.
