### 2017-01-31 785531b
merge database in edit.html
