+++ app/js/common.js
@@ -3,6 +3,16 @@
  * https://www.elabftw.net
  */
 
+function checkCookiesEnabled() {
+    var cookieEnabled = (navigator.cookieEnabled) ? true : false;
+    if (typeof navigator.cookieEnabled == "undefined" && !cookieEnabled) {
+        document.cookie="testcookie";
+        cookieEnabled = (document.cookie.indexOf("testcookie") != -1) ? true : false;
+    }
+
+    return (cookieEnabled);
+}
+
 // for editXP/DB, ctrl-shift-D will add the date
 function addDateOnCursor() {
     var todayDate = new Date();
