### 2017-02-21 0285605
remove sensiolab badge
