### 2017-02-27 elabftw/elabftw#349 @JohnGriffiths
can't install on ubuntu
# Description of the problem

I can't get the software to install and/or begin post-install configuration properly.

I follow in GNU linux installation instructions, but after installing with `elabctl install` and then typing `elabctl start`, I try to go to the  `<ip address>/install` page as described. But the browser just hangs.

Any idea how to fix this?

# Informations

* Elabftw version (visible in Sysadmin panel) :          ? (latest on day of posting)
* Installation method (git, docker or zip archive) :    docker
* Operating system + version :                                  ubuntu
* Browser used (firefox/chrome) :                              chrome & firefox

For git/zip installation method:

* PHP version:
* MySQL version:
* Webserver (apache/nginx) + version :
