### 2017-02-27 #146 @JohnGriffiths
Trying to but so far haven't been able to install

https://github.com/elabftw/elabftw/issues/349
