### 2017-02-27 elabftw/elabftw#349 @NicolasCARPi
Hello,

As explained in the documentation, it might take between 5 seconds and 3 minutes to start at first.

~~~bash
docker logs -f elabftw
~~~
will show you where it's at.
