### 2017-02-27 elabftw/elabftw#349 @JohnGriffiths
I think it's finished installing.

This is what `docker logs -f elabftw` shows:

>root@vm-MCIDS1:/home/jgriffiths/Desktop# docker logs -f elabftw
Generating a 4096 bit RSA private key
......................................................++
...........................++
writing new private key to '/etc/nginx/certs/server.key'

>Generating DH parameters, 2048 bit long safe prime, generator 2
This is going to take a long time

>......................................................................................................................................................................(LOTS OF DOTS...)

......................................................................................................................................................................

>2017-02-27 20:09:59,526 CRIT Supervisor running as root (no user in config file)
2017-02-27 20:09:59,546 INFO RPC interface 'supervisor' initialized
2017-02-27 20:09:59,546 CRIT Server 'unix_http_server' running without any HTTP authentication checking
2017-02-27 20:09:59,547 INFO supervisord started with pid 41
2017-02-27 20:10:00,552 INFO spawned: 'nginx' with pid 44
2017-02-27 20:10:00,556 INFO spawned: 'php7-fpm' with pid 45
2017-02-27 20:10:01,615 INFO success: nginx entered RUNNING state, process has stayed up for > than 1 seconds (startsecs)
2017-02-27 20:10:01,616 INFO success: php7-fpm entered RUNNING state, process has stayed up for > than 1 seconds (startsecs)

...but when I navigate to `https://192.133.45.196/install` in the browser, it just hangs.
