### 2017-02-27 elabftw/elabftw#349 @NicolasCARPi
The logs are correct. Are you using Firefox? I know I had some issues with Firefox, but not Chrome/Chromium. FF sometimes takes a long time to load the page, and the second one, too, but after that it's all fine. I think there is a bug in Firefox but didn't have time to track it down yet.
