### 2017-02-28 elabftw/elabftw#349 @JohnGriffiths
I've tried chrome, firefox, chromium. None of them work.
