### 2017-02-28 elabftw/elabftw#349 @NicolasCARPi
Well that's weird.

From the server:

~~~bash
curl -kL https://localhost
~~~
Should return the html of the login page.

If that works it means your issue is with your firewall ;)
