### 2017-03-09 elabftw/elabftw#349 @NicolasCARPi
@JohnGriffiths did you manage to get it working?
