### 2017-03-09 elabftw/elabftw#349 @JohnGriffiths
Looks like the error is `(113 "Host is unreachable")`

```
root@cpac:/home/jgriffiths# curl -kL https://localhost

<!DOCTYPE HTML>
<html>
<head>
<meta http-equiv="Content-type" content="text/html;charset=UTF-8" />
<meta http-equiv="X-UA-Compatible" content="IE=edge">
<meta name="viewport" content="width=device-width, initial-scale=1">
<meta name="author" content="Nicolas CARPi" />
<meta name='referrer' content='origin'>
<link rel="icon" type="image/ico" href="../app/img/favicon.ico" />
<title>eLabFTW - INSTALL</title>

<link rel="stylesheet" media="all" href="../app/css/elabftw.min.css" />

<script src="../app/js/elabftw.min.js"></script>
</head>

<body>
<section id="container" class='container'>
<section id='real_container'>
<center><img src='../app/img/logo.png' alt='elabftw' title='elabftw' /></center>
<h2>Welcome to the install of eLabFTW</h2>

<div class='alert alert-danger'><span class='glyphicon glyphicon-exclamation-sign' aria-hidden='true'></span><a href='#' class='close' data-dismiss='alert'>&times</a> SQLSTATE[HY000] [2003] Can't connect to MySQL server on 'mysql' (113 "Host is unreachable")</div></section></section></body></html>root@cpac:/home/jgriffiths#
```

So you think this is a firewall thing?
