### 2017-03-09 elabftw/elabftw#349 @NicolasCARPi
Hello,

Here the problem is that the "mysql" container cannot be reached by the "elabftw" container. This is why it was hanging, it was trying to connect to mysql but didn't manage to, and there is probably a timeout.

You didn't specify the Ubuntu version you are using. And it has its importance, because Docker works much better with 16.04, altough it worked fine with 14.04.

Is the mysql container running? It might not be running if for instance you ran out of memory or some other error like this.

Use:

~~~bash
docker ps
~~~
to see if it's running. And:
~~~bash
docker logs mysql
~~~
to see if you have an error.
If the last command ends with "[Note] mysqld: ready for connections.", then the issue is the communication between the two containers, and that's a Docker issue. It would help if you can provide the Docker version, too (`docker version`).

Also, if you just want to test it for JOSS, I can provide an image for you to test, already installed (I have free credits on DigitalOcean), it might be easier :) Or you can go to https://demo.elabftw.net to test it!
