### 2017-03-15 e9b9af9
fix display of images uploaded as album with fancybox
