### 2017-03-26 elabftw/elabftw#349 @JohnGriffiths
Ok how to proceed with digitalocean option for purposes of the review?
