### 2017-03-26 elabftw/elabftw#349 @NicolasCARPi
@JohnGriffiths Please check your gmail inbox, I sent you instructions on how to get to your instance :)
