### 2017-03-27 946fac8
minor css fixes
