+++ Gruntfile.js
@@ -22,7 +22,7 @@ module.exports = function(grunt) {
               'bower_components/bootstrap/js/button.js',
               'bower_components/bootstrap/js/dropdown.js',
               'bower_components/colorpicker/jquery.colorpicker.js',
-              'bower_components/fancybox/source/jquery.fancybox.js',
+              'bower_components/fancybox/dist/jquery.fancybox.js',
               'bower_components/jeditable/jquery.jeditable.js',
               'bower_components/jquery.complexify/jquery.complexify.js',
               'bower_components/jquery.complexify/jquery.complexify.banlist.js',
@@ -66,7 +66,7 @@ module.exports = function(grunt) {
               'bower_components/fullcalendar/dist/fullcalendar.css',
               'bower_components/bootstrap/dist/css/bootstrap.css',
               'bower_components/colorpicker/jquery.colorpicker.css',
-              'bower_components/fancybox/source/jquery.fancybox.css',
+              'bower_components/fancybox/dist/jquery.fancybox.css',
               'bower_components/jquery-ui/themes/smoothness/jquery-ui.css',
               'app/css/main.css'],
 
