### 2017-04-04 668275c
add possiblity to insert code with highlight. fix #375
add codesample plugin in tinymce and prism.js/.css
