### 2017-04-05 #146 @NicolasCARPi
@JohnGriffiths Now that you have access to your own install, did you manage to progress on the review?

Cheers,
~Nico
