### 2017-04-05 #146 @JohnGriffiths
@cMadan I've been through the cloud instance and the functionality all seems fine. I've updated all the tick boxes above. The only thing I can't give a tick for is installation, since as you know I wasn't able to install successfully on linux.
