### 2017-04-05 #146 @NicolasCARPi
That's really too bad that you didn't manage to install. I'll just leave this here:

![2017-04-05-231703_789x169_scrot](https://cloud.githubusercontent.com/assets/3043706/24727409/2b334c56-1a56-11e7-8085-cabb8bbebd4e.png)
