### 2017-04-12 #146 @cMadan
@JohnGriffiths - thanks for your work in reviewing this! I know you spent a good deal of time trying to evaluate it and I really appreciate it

@NicolasCARPi - I am sufficiently satisfied that others have been able to install eLabFTW, and your Github issues are quite active, that I believe that others will be able to set-up the software. If you can provide the DOI to an archive of the current version of the software, I am happy to accept your paper at JOSS.
