### 2017-04-12 #146 @NicolasCARPi
@cMadan Great news :) What do you mean by "provide a DOI to an archive"?

This is a link to the latest version:
https://github.com/elabftw/elabftw/archive/1.5.6.tar.gz
