### 2017-04-12 #146 @cMadan
@NicolasCARPi - Basically, we require that the archive be stored on a service such as Zenodo or figshare, please see the end of the author guidelines section here (http://joss.theoj.org/about#author_guidelines)
