### 2017-04-14 #146 @arfon
@NicolasCARPi - please could you remove the "Links" section as this breaks the paper compilation with pandoc.

```
# Links

- [Official website](https://www.elabftw.net)
- [Documentation](https://elabftw.readthedocs.io)
- [Live demo](https://demo.elabftw.net)
- [GitHub repository](https://github.com/elabftw/elabftw)
```

We already link to the GitHub repository in the compiled PDF so this one can be removed and the other links should be cited in the text in the main `Summary` section (and then the references in a bibtex file). You can see a recent example of how to do this here: https://github.com/ropensci/getCRUCLdata/tree/master/inst/paper
