### 2017-04-14 f5e25de
add paper.bib and remove Links section from paper.md.
See openjournals/joss-reviews#146
