### 2017-04-14 #146 @NicolasCARPi
@arfon I made the requested changes (see commit above). Cheers. Nico.
